% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.r
\name{add_intron_pos_index}
\alias{add_intron_pos_index}
\title{Add intron position index to graph list}
\usage{
add_intron_pos_index(t_igraph_list, t_intron_pos_mat_fr)
}
\arguments{
\item{t_igraph_list}{List of graph objects}

\item{t_intron_pos_mat_fr}{Data frame of intron positions}
}
\value{
Updated graph list with intron position index
}
\description{
Add intron position index to graph list
}
