% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_building.r
\name{build_adjacency_matrices}
\alias{build_adjacency_matrices}
\title{Build adjacency matrices for transcripts}
\usage{
build_adjacency_matrices(
  t_iso_final,
  t_iso_slow_sumary,
  intron_pair_cov_threshold = 0.95,
  t_read_count_threshold = 0
)
}
\arguments{
\item{t_iso_final}{Data frame of intron splicing order pairs}

\item{t_iso_slow_sumary}{Data frame with isoform summary statistics}

\item{intron_pair_cov_threshold}{Coverage threshold for intron pairs}

\item{t_read_count_threshold}{Read count threshold}
}
\value{
List of adjacency matrices for each transcript
}
\description{
Build adjacency matrices for transcripts
}
