% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.r
\name{build_iso_object}
\alias{build_iso_object}
\title{Build isoform object from intron splicing order files}
\usage{
build_iso_object(files_all, intron_anno, trans_exp_file = "")
}
\arguments{
\item{files_all}{Vector of file paths containing intron splicing order data}

\item{intron_anno}{Data frame of intron annotations}

\item{trans_exp_file}{Optional file with expressed transcript IDs}
}
\value{
Data frame of intron splicing order pairs
}
\description{
Build isoform object from intron splicing order files
}
