% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cal_intron_pair_cov}
\alias{cal_intron_pair_cov}
\title{Calculate intron pair coverage}
\usage{
cal_intron_pair_cov(t_mat)
}
\arguments{
\item{t_mat}{Adjacency matrix}
}
\value{
Percentage coverage of intron pairs
}
\description{
Calculate intron pair coverage
}
