% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_pipeline.r
\name{calculate_heterogeneity}
\alias{calculate_heterogeneity}
\title{Calculate heterogeneity of splicing patterns}
\usage{
calculate_heterogeneity(t_igraph_list, t_alpha)
}
\arguments{
\item{t_igraph_list}{List of graph objects}

\item{t_alpha}{Smoothing parameter}
}
\value{
Updated graph list with heterogeneity measures
}
\description{
Calculate heterogeneity of splicing patterns
}
