% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_pipeline.r
\name{calculate_most_likely_order}
\alias{calculate_most_likely_order}
\title{Calculate most likely order for all transcripts}
\usage{
calculate_most_likely_order(t_igraph_list, t_alpha = 0.1)
}
\arguments{
\item{t_igraph_list}{List of graph objects}

\item{t_alpha}{Smoothing parameter}
}
\value{
Updated graph list with most likely orders
}
\description{
Calculate most likely order for all transcripts
}
