% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlp_algorithms.r
\name{calp2}
\alias{calp2}
\title{Calculate probability of intron order}
\usage{
calp2(t_adj_mat, t_intron_order, t_alpha_v)
}
\arguments{
\item{t_adj_mat}{Adjacency matrix}

\item{t_intron_order}{Vector of intron orders}

\item{t_alpha_v}{Smoothing parameter}
}
\value{
Log probability of the order
}
\description{
Calculate probability of intron order
}
