% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_iso.R
\name{check_file_format}
\alias{check_file_format}
\title{检查文件格式}
\usage{
check_file_format(file_path, file_type = "bed")
}
\arguments{
\item{file_path}{文件路径}

\item{file_type}{文件类型（"bed"或"bam"）}
}
\description{
检查文件格式
}
