% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.r
\name{cluster_introns}
\alias{cluster_introns}
\title{Cluster introns based on splicing patterns}
\usage{
cluster_introns(t_igraph_list, t_alpha = 0.05)
}
\arguments{
\item{t_igraph_list}{List of graph objects}

\item{t_alpha}{Smoothing parameter for clustering}
}
\value{
Updated graph list with cluster membership
}
\description{
Cluster introns based on splicing patterns
}
