% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.r
\name{cluster_introns_matrix}
\alias{cluster_introns_matrix}
\title{Cluster introns from adjacency matrix}
\usage{
cluster_introns_matrix(tt_adj_mat, t_alpha_v = 0.05)
}
\arguments{
\item{tt_adj_mat}{Adjacency matrix}

\item{t_alpha_v}{Smoothing parameter}
}
\value{
Vector of cluster assignments
}
\description{
Cluster introns from adjacency matrix
}
