% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_results.r
\name{draw_mlo_plot}
\alias{draw_mlo_plot}
\title{Draw MLO plot for transcript order visualization}
\usage{
draw_mlo_plot(
  tt_igraph_list,
  output_path,
  all_alt_se_introns = NULL,
  all_alt_a5_introns = NULL,
  all_alt_a3_introns = NULL,
  all_alt_ri_introns = NULL
)
}
\arguments{
\item{tt_igraph_list}{List of graph objects}

\item{output_path}{Path to output PDF file}

\item{all_alt_se_introns}{Vector of SE intron positions}

\item{all_alt_a5_introns}{Vector of A5 intron positions}

\item{all_alt_a3_introns}{Vector of A3 intron positions}

\item{all_alt_ri_introns}{Vector of RI intron positions}
}
\value{
NULL (creates PDF file)
}
\description{
Draw MLO plot for transcript order visualization
}
