% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_results.r
\name{draw_mol_table_plot}
\alias{draw_mol_table_plot}
\title{Draw probability table plot for order comparisons}
\usage{
draw_mol_table_plot(t_igraph_list, output_path)
}
\arguments{
\item{t_igraph_list}{List of graph objects}

\item{output_path}{Path to output PDF file}
}
\value{
NULL (creates PDF file)
}
\description{
Draw probability table plot for order comparisons
}
