% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.r
\name{extract_introns_from_bed}
\alias{extract_introns_from_bed}
\title{Extract exons and introns from BED file}
\usage{
extract_introns_from_bed(t_bed_file, m_trim_trans_id_by_dot = TRUE)
}
\arguments{
\item{t_bed_file}{Path to BED file}

\item{m_trim_trans_id_by_dot}{Whether to trim transcript ID by dot (default: TRUE)}
}
\value{
List containing intron and exon data frames
}
\description{
Extract exons and introns from BED file
}
