% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlp_algorithms.r
\name{find_path_global}
\alias{find_path_global}
\title{Find most likely intron splicing order}
\usage{
find_path_global(t_adj_mat, t_alpha_v = 0.1, is_verbose = FALSE)
}
\arguments{
\item{t_adj_mat}{Adjacency matrix}

\item{t_alpha_v}{Smoothing parameter}

\item{is_verbose}{Whether to print verbose output}
}
\value{
List containing best order and statistics
}
\description{
Find most likely intron splicing order
}
