% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_iso.R
\name{get_bam_data_for_transcript}
\alias{get_bam_data_for_transcript}
\title{获取转录本的BAM数据}
\usage{
get_bam_data_for_transcript(bam_file, transcript, padding = 1000, min_mapq = 1)
}
\arguments{
\item{bam_file}{BAM文件路径}

\item{transcript}{转录本信息}

\item{padding}{区域扩展大小}

\item{min_mapq}{最小MAPQ值}
}
\value{
BAM数据列表
}
\description{
获取转录本的BAM数据
}
