% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.r
\name{get_iso_summary}
\alias{get_iso_summary}
\title{Calculate isoform summary statistics}
\usage{
get_iso_summary(t_iso_final, t_anno_intron)
}
\arguments{
\item{t_iso_final}{Data frame of intron splicing order pairs}

\item{t_anno_intron}{Data frame of intron annotations}
}
\value{
Data frame with isoform summary statistics
}
\description{
Calculate isoform summary statistics
}
