% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_iso.R
\name{getIsoFromBam}
\alias{getIsoFromBam}
\title{从BAM文件生成iso_file的主函数}
\usage{
getIsoFromBam(
  bed_file,
  bam_file,
  output_file,
  intron_flank_threshold = 90,
  consider_exon_in_intron = TRUE,
  unique_intron = FALSE,
  min_mapq = 1,
  n_threads = 1,
  padding = 1000
)
}
\arguments{
\item{bed_file}{BED文件路径}

\item{bam_file}{BAM文件路径}

\item{output_file}{输出文件路径}

\item{intron_flank_threshold}{内含子侧翼阈值}

\item{consider_exon_in_intron}{是否考虑外显子在内含子中}

\item{unique_intron}{是否只使用唯一内含子}

\item{min_mapq}{最小MAPQ值}

\item{n_threads}{线程数}

\item{padding}{区域扩展大小}
}
\value{
分析结果列表
}
\description{
从BAM文件生成iso_file的主函数
}
