% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{if_smaller_f}
\alias{if_smaller_f}
\title{Check if region1 is smaller than region2}
\usage{
if_smaller_f(region1, region2)
}
\arguments{
\item{region1}{First genomic region}

\item{region2}{Second genomic region}
}
\value{
Logical vector
}
\description{
Check if region1 is smaller than region2
}
