% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlp_algorithms.r
\name{lp_kenemy}
\alias{lp_kenemy}
\title{Linear programming approach for Kenemy problem}
\usage{
lp_kenemy(t_adj_mat, t_alpha_v, verbose_hill = FALSE)
}
\arguments{
\item{t_adj_mat}{Adjacency matrix}

\item{t_alpha_v}{Smoothing parameter}

\item{verbose_hill}{Whether to print verbose output}
}
\value{
List containing best order
}
\description{
Linear programming approach for Kenemy problem
}
