% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_results.r
\name{output_mlo_results}
\alias{output_mlo_results}
\title{Output most likely order results}
\usage{
output_mlo_results(
  t_igraph_list,
  t_gene_trans_id_map,
  t_intron_pos_mat_fr,
  output_file,
  trim_trans_id_by_dot
)
}
\arguments{
\item{t_igraph_list}{List of graph objects}

\item{t_gene_trans_id_map}{Optional gene-transcript mapping}

\item{t_intron_pos_mat_fr}{Intron position data frame}

\item{output_file}{Path to output file}

\item{trim_trans_id_by_dot}{Whether to trim transcript ID by dot}
}
\value{
Updated graph list
}
\description{
Output most likely order results
}
