% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_iso.R
\name{prepare_transcript_data}
\alias{prepare_transcript_data}
\title{从BED文件准备转录本数据}
\usage{
prepare_transcript_data(bed_file, unique_intron = FALSE)
}
\arguments{
\item{bed_file}{BED文件路径}

\item{unique_intron}{是否只使用唯一内含子}
}
\value{
转录本列表
}
\description{
从BED文件准备转录本数据
}
